starqueueserver/requirements.txt
aiofiles==0.5.0
aiomysql==0.0.20
arrow==0.15.8
asyncpg==0.20.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
cryptography==3.0
defusedxml==0.6.0
docutils==0.15.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httptools==0.1.1
httpx==0.11.1
humanfriendly==8.2
hyperframe==5.2.0
idna==2.10
jmespath==0.10.0
multidict==4.7.6
mysql-connector-python==8.0.21
protobuf==3.13.0
pycparser==2.20
pydantic==1.6.1
PyMySQL==0.9.2
pyodbc==4.0.30
PySnooper==0.4.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-multipart==0.0.5
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
starlette==0.13.6
termcolor==1.1.0
ujson==3.0.0
urllib3==1.25.10
uvicorn==0.11.6
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
